the last version

+++ vm/visualization/musica.c
@@ -16,7 +16,8 @@ void		define_music(t_core *a)
 {
 	static char *tmp;
 
-	tmp = "afplay /Users/omaslova/correwar_git/vm/music/";
+	VIS->music.is_start = true;
+	tmp = "afplay /./vm/music/";
 	VIS->music.m_fight = ft_strjoin(tmp, "epic.mp3 &> /dev/null &");
 	VIS->music.magic = ft_strjoin(tmp, "boogie.mp3 &> /dev/null &");
 }
@@ -24,16 +25,13 @@ void		define_music(t_core *a)
 void		play_music(t_core *a, int what_now)
 {
 	if (what_now == MUS_FIGHT)
-	{
 		system(VIS->music.m_fight);
-	}
 	else if (what_now == MUS_MAGIC)
 	{
-		system("pkill afplay");
+		if (VIS->music.is_start)
+			system("pkill afplay");
 		system(VIS->music.magic);
 	}
 	else if (what_now == MUS_END)
-	{
 		system("pkill afplay");
-	}
 }
